+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.2.2
-* BREAKING CHANGE: `ScanResult` now returns a `BluetoothDevice`
+* BREAKING CHANGES
+* `startScan` renamed to `scan`
+* `ScanResult` now returns a `BluetoothDevice`
+* `connect()` now takes a `BluetoothDevice` and returns Stream<BluetoothDeviceState>
+* Automatic disconnect on deviceConnection.cancel()
 
 ## 0.2.1
 * BREAKING CHANGE: removed `stopScan` from API, use `scanSubscription.cancel()` instead